src/main/java/ladder/domain/Name.java
package ladder.domain;

import java.util.Objects;

public class Name {
    private static final int MAX_NAME_LENGTH = 5;

    private final String name;

    private Name(String name) {
        this.name = name;
    }

    public static Name from(final String name) {
        if (name.length() > MAX_NAME_LENGTH) {
            throw new RuntimeException(String.format("이름은 최대 %d자 까지 가능합니다.", MAX_NAME_LENGTH));
        }
        return new Name(name);
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        Name name1 = (Name) o;
        return Objects.equals(name, name1.name);
    }

    @Override
    public int hashCode() {
        return Objects.hash(name);
    }
}
